refactor: fix SRP issue with user management
