+++ src/shared/locales/en.js
@@ -1,3 +1,1 @@
-/*eslint-disable*/ module.exports = {
-  messages: JSON.parse('{"J/hVSQ":[["0"]]}'),
-};
+/*eslint-disable*/module.exports={messages:JSON.parse("{\"J/hVSQ\":[[\"0\"]]}")};
